Move frechet mean code to frechet_mean.py
